requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.6
Django==4.2.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
Markdown==3.4.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1

django-filter~=23.2